DMT5 signal trading page & added notice prop for side tab components
