+++ src/pages/dmt5-trading-signals/_subscription.js
@@ -1,9 +1,10 @@
 import React from 'react'
 import styled from 'styled-components'
 import { graphql, useStaticQuery } from 'gatsby'
-import { Text, Header, QueryImage } from 'components/elements'
-import { localize } from 'components/localization'
+import { Text, Header, QueryImage, LocalizedLinkText } from 'components/elements'
+import { localize, Localize } from 'components/localization'
 import { Container, SectionContainer, Flex } from 'components/containers'
+import { LinkButton } from 'components/form'
 import device from 'themes/device'
 
 const query = graphql`
@@ -47,7 +48,7 @@ const StyledHeader = styled(Header)`
 `
 
 const StyledText = styled(Text)`
-    font-size: 2.4rem;
+    font-size: 1.6rem;
     line-height: 3.6rem;
     max-width: 97.6rem;
     text-align: center;
@@ -62,6 +63,8 @@ const StyledText = styled(Text)`
 `
 
 const CardContainer = styled(Flex)`
+    margin-top: 4rem;
+
     @media ${device.laptopM} {
         max-width: 58rem;
         flex-wrap: wrap;
@@ -71,7 +74,7 @@ const Card = styled(Flex)`
     flex-direction: column;
     width: 39.1rem;
     height: auto;
-    margin: 6.3rem 2.5rem 3.2rem 0;
+    margin-right: 2.5rem;
     justify-content: flex-start;
 
     :last-child {
@@ -95,7 +98,7 @@ const Card = styled(Flex)`
     }
 `
 const StyledCardText = styled(Text)`
-    font-size: 24px;
+    font-size: 1.6rem;
     text-align: left;
     margin-top: 16px;
 
@@ -112,27 +115,30 @@ const Column = styled.div`
     width: 10rem;
 
     :first-child {
-        flex-basis: 10%;
+        flex-basis: 6%;
     }
 `
-
 const Row = styled.div`
     display: flex;
     flex-direction: row;
     width: 100%;
 `
+const NoteContainer = styled(Flex)`
+    flex-direction: column;
+    margin-top: 6.4rem;
+`
 
-// const DMT5Button = styled(LinkButton)`
-//     margin-top: 4rem;
-//     height: 40px;
-//     padding: 10px 16px;
-//     font-size: 14px;
-//     white-space: nowrap;
+const DMT5Button = styled(LinkButton)`
+    margin-top: 4rem;
+    height: 40px;
+    padding: 10px 16px;
+    font-size: 14px;
+    white-space: nowrap;
 
-//     @media ${device.tabletL} {
-//         margin-top: 32px;
-//     }
-// `
+    @media ${device.tabletL} {
+        margin-top: 32px;
+    }
+`
 
 const Subscription = () => {
     const data = useStaticQuery(query)
@@ -160,9 +166,10 @@ const Subscription = () => {
                             </Column>
                             <Column>
                                 <StyledCardText>
-                                    {localize(
-                                        'From your Deriv MT5 trading terminal, click on the Signals tab.',
-                                    )}
+                                    <Localize
+                                        translate_text="From your Deriv MT5 trading terminal, click on the <0>Signals</0> tab."
+                                        components={[<strong key={0} />]}
+                                    />
                                 </StyledCardText>
                             </Column>
                         </Row>
@@ -179,9 +186,10 @@ const Subscription = () => {
                             </Column>
                             <Column>
                                 <StyledCardText>
-                                    {localize(
-                                        'Click on the My Statistics tab and select your signal provider. ',
-                                    )}
+                                    <Localize
+                                        translate_text="Click on the <0>My Statistics</0> tab and select your signal provider."
+                                        components={[<strong key={0} />]}
+                                    />
                                 </StyledCardText>
                             </Column>
                         </Row>
@@ -198,14 +206,82 @@ const Subscription = () => {
                             </Column>
                             <Column>
                                 <StyledCardText>
-                                    {localize(
-                                        'Click the Renew button to renew the service or the Unsubscribe button to cancel the subscription.',
-                                    )}
+                                    <Localize
+                                        translate_text="Click the <0>Renew</0> button to renew the service or the <0>Unsubscribe</0> button to cancel the subscription."
+                                        components={[<strong key={0} />]}
+                                    />
                                 </StyledCardText>
                             </Column>
                         </Row>
                     </Card>
                 </CardContainer>
+                <NoteContainer>
+                    <StyledCardText>
+                        <Localize translate_text="<0>Note:</0>" components={[<strong key={0} />]} />
+                    </StyledCardText>
+                    <StyledCardText>
+                        <Localize
+                            translate_text="1. You must have an MQL5 community account to subscribe to trading signals. If you don't have an account yet, please go to <0>MQL5.com</0> to register."
+                            components={[
+                                <LocalizedLinkText
+                                    external="true"
+                                    to="https://www.mql5.com/en/auth_register"
+                                    color="red"
+                                    size={16}
+                                    key={0}
+                                />,
+                            ]}
+                        />
+                    </StyledCardText>
+                    <StyledCardText>
+                        <Localize
+                            translate_text="2. You'll need to <0>fund your MQL5 account</0> to subscribe to paid signals."
+                            components={[
+                                <LocalizedLinkText
+                                    external="true"
+                                    to="https://www.mql5.com/en/articles/302#deposit"
+                                    color="red"
+                                    size={16}
+                                    key={0}
+                                />,
+                            ]}
+                        />
+                    </StyledCardText>
+                    <StyledCardText>
+                        {localize(
+                            '3. You can only subscribe to one signal provider with one DMT5 account at any given time. You can use your signal on up to 3 computers.',
+                        )}
+                    </StyledCardText>
+                    <StyledCardText>
+                        {localize(
+                            "4. You won't be able to trade manually with the same DMT5 account once you've subscribed to a signal.",
+                        )}
+                    </StyledCardText>
+                    <StyledCardText>
+                        <Localize
+                            translate_text="Have more questions? Check out the <0>FAQ section</0> on the MQL5 website."
+                            components={[
+                                <LocalizedLinkText
+                                    external="true"
+                                    to="https://www.mql5.com/en/forum/10773#q17"
+                                    color="red"
+                                    size={16}
+                                    key={0}
+                                />,
+                            ]}
+                        />
+                    </StyledCardText>
+                </NoteContainer>
+                <DMT5Button
+                    secondary="true"
+                    external="true"
+                    is_deriv_app_link
+                    to="/mt5"
+                    target="_blank"
+                    rel="noopener noreferrer"
+                >
+                    {localize('Go to your DMT5 terminal')}
+                </DMT5Button>
             </StyledContainer>
         </StyledSectionContainer>
     )
