+++ src/pages/dmt5-trading-signals/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
-import { Hero, SmallContainer, StyledHeader } from './components/_style'
-import HowTo from './components/_how-to'
+import { Hero, SmallContainer, StyledHeader } from './_style'
+import HowTo from './_how-to'
 import { Signal } from './_signal'
 import SignalSteps from './_signal-steps'
 import Subscription from './_subscription'
